# aula06/doc/openapi (1).yaml
# shift + alt+ p
openapi: 3.0.3
info:
  title: Projeto do Filmes DS2M - SENAI Jandira 
  description: API destinada para controlar Filmes
  termsOfService: http://swagger.io/terms/
  version: 1.0.4.26
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://
paths:
  /v1/senai/locadora/filme:
    post:
      tags:
        - filme
      summary: Adiciona um novo Filme
      description: Adiciona um novo Filme no Banco de Dados
      requestBody:
        description: JSON contendo os dados do novo Filme a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Successful operation
        '400':
          description: Invalid input
        '500':
          description: Error Model
    get:
      tags:
        - filme
      summary: Lista todos os filmes
      description: Listagem de todos os filmes registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filme'          
           
        '404':
          description: Not Found
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - filme
      summary: Busca um filme pelo ID
      description: Retorna um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para retornar o filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'          
        '400':
          description: Invalid ID supplied
        '404':
          description: Filme not found
    put:
      tags:
        - filme
      summary: Atualizar um filme existente pelo ID
      description: Localizar e atualizar um filme buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para localizar e atualizar o filme
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Daados para atualizar o filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
    delete:
      tags:
        - filme
      summary: Excluir um filme pelo Id
      description: Localizar e excluir um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para localizar o Filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
  

components:
  schemas:
    Filme:
      required:
        - nome
        - sinopse
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        sinopse:
          type: string
        data_lancamento:
          type: string
        duracao:
          type: string
        valor:
          type: integer
        avaliacao:
          type: integer
        capa:
          type: string
  requestBodies:
    Filme:
      description: Objeto que deve ser enviado no Body das requisicoes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Filme'
    